+++ features/openMultiplyVault/openMultiplyVault.ts
@@ -1,24 +1,24 @@
 import { BigNumber } from 'bignumber.js'
 import { maxUint256 } from 'blockchain/calls/erc20'
 import { createIlkDataChange$, IlkData } from 'blockchain/ilks'
-import { ContextConnected } from 'blockchain/network'
+import { ContextConnected, every5Seconds$ } from 'blockchain/network'
 import { TxHelpers } from 'components/AppContext'
-import { ExchangeAction, Quote } from 'features/exchange/exchange'
+import { Quote } from 'features/exchange/exchange'
 import { BalanceInfo, balanceInfoChange$ } from 'features/shared/balanceInfo'
 import { PriceInfo, priceInfoChange$ } from 'features/shared/priceInfo'
 import { curry } from 'lodash'
-import { combineLatest, EMPTY, iif, merge, Observable, of, Subject, throwError } from 'rxjs'
+import { combineLatest, iif, merge, Observable, of, Subject, throwError } from 'rxjs'
 import {
   debounceTime,
   distinctUntilChanged,
   first,
   map,
-  mergeMap,
   scan,
   shareReplay,
   switchMap,
 } from 'rxjs/operators'
 
+import { applyExchange, applyQuote, ExchangeQuoteChanges, SLIPPAGE } from './openMultiplyQuote'
 import { applyOpenVaultAllowance, OpenVaultAllowanceChange } from './openMultiplyVaultAllowances'
 import {
   applyOpenVaultCalculations,
@@ -71,39 +71,6 @@ function applyOpenVaultInjectedOverride(
       ...change.stateToOverride,
     }
   }
-  return state
-}
-
-type ExchangeQuoteSuccessChange = {
-  kind: 'quote'
-  quote: Quote
-}
-
-type ExchangeQuoteFailureChange = {
-  kind: 'quoteError'
-}
-
-type ExchangeQuoteResetChange = {
-  kind: 'quoteReset'
-}
-type ExchangeQuoteChanges =
-  | ExchangeQuoteSuccessChange
-  | ExchangeQuoteFailureChange
-  | ExchangeQuoteResetChange
-
-function applyExchange(change: OpenMultiplyVaultChange, state: OpenMultiplyVaultState) {
-  if (change.kind === 'quote') {
-    return {
-      ...state,
-      quote: change.quote,
-    }
-  }
-
-  if (change.kind === 'quoteReset') {
-    const { quote: _quote, ...rest } = state
-    return rest
-  }
-
   return state
 }
 
@@ -304,59 +271,6 @@ export const defaultMutableOpenVaultState: MutableOpenMultiplyVaultState = {
   allowanceAmount: maxUint256,
 }
 
-const SLIPPAGE = new BigNumber(0.05)
-
-function applyQuote(
-  exchangeQuote$: (
-    token: string,
-    slippage: BigNumber,
-    amount: BigNumber,
-    action: ExchangeAction,
-  ) => Observable<Quote>,
-  state$: Observable<OpenMultiplyVaultState>,
-): Observable<OpenMultiplyVaultChange> {
-  return state$.pipe(
-    debounceTime(500),
-    distinctUntilChanged(
-      (s1, s2) =>
-        s1.token === s2.token &&
-        s1.slippage.eq(s2.slippage) &&
-        s1.buyingCollateral.eq(s2.buyingCollateral),
-    ),
-    switchMap((state) =>
-      iif(
-        () => state.buyingCollateral.gt(0),
-
-        exchangeQuote$(state.token, state.slippage, state.buyingCollateral, 'BUY').pipe(
-          mergeMap((quote) =>
-            quote.status === 'SUCCESS' ? of({ kind: 'quote', quote }) : of({ kind: 'quoteError' }),
-          ),
-        ),
-        state.quote === undefined ? EMPTY : of({ kind: 'quoteReset' }),
-      ),
-    ),
-  )
-}
-
-function applyQuote_(
-  exchangeQuote$: (
-    token: string,
-    slippage: BigNumber,
-    amount: BigNumber,
-    action: ExchangeAction,
-  ) => Observable<Quote>,
-  state: OpenMultiplyVaultState,
-): Observable<OpenMultiplyVaultChange> {
-  if (state.buyingCollateral.gt(0)) {
-    return exchangeQuote$(state.token, state.slippage, state.buyingCollateral, 'BUY').pipe(
-      mergeMap((quote) =>
-        quote.status === 'SUCCESS' ? of({ kind: 'quote', quote }) : of({ kind: 'quoteError' }),
-      ),
-    )
-  }
-  return state.quote === undefined ? EMPTY : of({ kind: 'quoteReset' })
-}
-
 export function createOpenMultiplyVault$(
   context$: Observable<ContextConnected>,
   txHelpers$: Observable<TxHelpers>,
@@ -432,7 +346,6 @@ export function createOpenMultiplyVault$(
 
                     const state$ = merge(change$, environmentChanges$).pipe(
                       scan(apply, initialState),
-
                       map(validateErrors),
                       map(validateWarnings),
                       map(curry(addTransitions)(txHelpers, connectedProxyAddress$, change)),
@@ -447,7 +360,9 @@ export function createOpenMultiplyVault$(
                             s1.slippage.eq(s2.slippage) &&
                             s1.buyingCollateral.eq(s2.buyingCollateral),
                         ),
-                        switchMap((state) => applyQuote_(exchangeQuote$, state)),
+                        switchMap((state) =>
+                          every5Seconds$.pipe(switchMap(() => applyQuote(exchangeQuote$, state))),
+                        ),
                       )
                       .subscribe(change)
